refactor: JSONデータ作成とファイル処理を別クラスに分離させた
Issue: #16
